src/main/java/com/qa/appadmin/pages/AppAdminDeleteConfirmationPage.java
package com.qa.appadmin.pages;

import java.io.IOException;

import org.openqa.selenium.WebElement;
import org.openqa.selenium.interactions.Actions;
import org.openqa.selenium.support.FindBy;
import org.openqa.selenium.support.PageFactory;

import com.qa.base.TestBase;

public class AppAdminDeleteConfirmationPage extends TestBase {
	@FindBy(xpath="//button[@type='submit']")
	WebElement DeleteButton;
	
	@FindBy(xpath="//a[@class='btn btn-default']")
	WebElement CancelButton;
	
	@FindBy(xpath="//*[@id=\"mainmenubutton\"]")
	WebElement mainMenuButton;
	
	@FindBy(xpath="//*[text()='Configuration']")
	WebElement configurationLink;
	
	@FindBy(xpath="//*[text()='Reports']")
	WebElement reportsLink ;
	
	@FindBy(xpath="//*[@id=\"dropdownMenu1\"]")
	WebElement userName;
	
	@FindBy(xpath="/html/body/header/div/div/div[3]/div/ul/li[1]/a")
	WebElement changePasswordLink ;
	
	@FindBy(xpath="/html/body/header/div/div/div[3]/div/ul/li[2]/a")
	WebElement logoutLink ;
	
	@FindBy(xpath="/html/body/section/h2")
	WebElement pageTitle ;
	
	//initializing page object
	public AppAdminDeleteConfirmationPage() throws IOException {
		PageFactory.initElements(driver, this);
	}
	
//Actions
	public String verifyStudiesPageTitle() {
		return driver.getTitle();
	}
	
	public boolean verifyCorrectUsername() {
		return userName.isDisplayed();
	}
	public AppAdminChangePasswordPage clickOnChangePasswordLink() throws IOException {
		userName.click();
		Actions act=new Actions(driver);
		act.click(userName).moveToElement(changePasswordLink);
		changePasswordLink.click();
		return new AppAdminChangePasswordPage();
	}
	
	public AppAdminHomePage clickOnLogoutLink() throws IOException {
		userName.click();
		Actions act=new Actions(driver);
		act.click(userName).moveToElement(logoutLink);
		logoutLink.click();
		return new AppAdminHomePage();
	}
	
	public AppAdminStudiesPage clickOnDeleteButton() throws IOException {
		DeleteButton.click();
		return new AppAdminStudiesPage();
		
	}
	public AppAdminStudiesPage clickOnCancelButton() throws IOException {
		CancelButton.click();
		return new AppAdminStudiesPage();
		
	}

	
}
